requirements.txt
ipython==8.12.3
ipywidgets==8.1.2
matplotlib==3.7.5
numpy==1.24.4
numpy==1.24.3
open3d==0.18.0
opencv_contrib_python==4.8.0.74
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==12.0.0
plotly==5.22.0
scikit_learn==1.3.2
scipy==1.16.3
skimage==0.0
tabulate==0.9.0
tqdm==4.65.2
trimesh==4.4.1
